+++ crates/query/src/plan/type_join/type_join_many/mod.rs
@@ -1,0 +1,7 @@
+mod children;
+mod compare;
+mod node;
+mod plan_node;
+
+pub use compare::{compare_json_values, resolve_nested_field};
+pub use node::TypeJoinMany;
